refactor(engine): split tool execution helpers
