+++ crates/tui/src/core/engine/tool_execution.rs
@@ -1,0 +1,197 @@
+//! Low-level tool execution helpers for the engine turn loop.
+//!
+//! This module keeps the mechanics of MCP dispatch, execution locking, and
+//! parallel-tool fanout out of `engine.rs`; the turn loop still owns planning,
+//! approval, and how tool results are written back into session state.
+
+use std::{fs::OpenOptions, io::Write};
+
+use super::*;
+
+pub(super) fn emit_tool_audit(event: serde_json::Value) {
+    let Some(path) = std::env::var_os("DEEPSEEK_TOOL_AUDIT_LOG") else {
+        return;
+    };
+    let line = match serde_json::to_string(&event) {
+        Ok(line) => line,
+        Err(_) => return,
+    };
+    let path = PathBuf::from(path);
+    if let Some(parent) = path.parent() {
+        let _ = std::fs::create_dir_all(parent);
+    }
+    if let Ok(mut file) = OpenOptions::new().create(true).append(true).open(path) {
+        let _ = writeln!(file, "{line}");
+    }
+}
+
+impl Engine {
+    pub(super) async fn execute_mcp_tool_with_pool(
+        pool: Arc<AsyncMutex<McpPool>>,
+        name: &str,
+        input: serde_json::Value,
+    ) -> Result<ToolResult, ToolError> {
+        let mut pool = pool.lock().await;
+        let result = pool
+            .call_tool(name, input)
+            .await
+            .map_err(|e| ToolError::execution_failed(format!("MCP tool failed: {e}")))?;
+        let content = serde_json::to_string_pretty(&result).unwrap_or_else(|_| result.to_string());
+        Ok(ToolResult::success(content))
+    }
+
+    pub(super) async fn execute_parallel_tool(
+        &mut self,
+        input: serde_json::Value,
+        tool_registry: Option<&crate::tools::ToolRegistry>,
+        tool_exec_lock: Arc<RwLock<()>>,
+    ) -> Result<ToolResult, ToolError> {
+        let calls = parse_parallel_tool_calls(&input)?;
+        let mcp_pool = if calls.iter().any(|(tool, _)| McpPool::is_mcp_tool(tool)) {
+            Some(self.ensure_mcp_pool().await?)
+        } else {
+            None
+        };
+        let Some(registry) = tool_registry else {
+            return Err(ToolError::not_available(
+                "tool registry unavailable for multi_tool_use.parallel",
+            ));
+        };
+
+        let mut tasks = FuturesUnordered::new();
+        for (tool_name, tool_input) in calls {
+            if tool_name == MULTI_TOOL_PARALLEL_NAME {
+                return Err(ToolError::invalid_input(
+                    "multi_tool_use.parallel cannot call itself",
+                ));
+            }
+            if McpPool::is_mcp_tool(&tool_name) {
+                if !mcp_tool_is_parallel_safe(&tool_name) {
+                    return Err(ToolError::invalid_input(format!(
+                        "Tool '{tool_name}' is an MCP tool and cannot run in parallel. \
+                         Allowed MCP tools: list_mcp_resources, list_mcp_resource_templates, \
+                         mcp_read_resource, read_mcp_resource, mcp_get_prompt."
+                    )));
+                }
+            } else {
+                let Some(spec) = registry.get(&tool_name) else {
+                    return Err(ToolError::not_available(format!(
+                        "tool '{tool_name}' is not registered"
+                    )));
+                };
+                if !spec.is_read_only() {
+                    return Err(ToolError::invalid_input(format!(
+                        "Tool '{tool_name}' is not read-only and cannot run in parallel"
+                    )));
+                }
+                if spec.approval_requirement() != ApprovalRequirement::Auto {
+                    return Err(ToolError::invalid_input(format!(
+                        "Tool '{tool_name}' requires approval and cannot run in parallel"
+                    )));
+                }
+                if !spec.supports_parallel() {
+                    return Err(ToolError::invalid_input(format!(
+                        "Tool '{tool_name}' does not support parallel execution"
+                    )));
+                }
+            }
+
+            let registry_ref = registry;
+            let lock = tool_exec_lock.clone();
+            let tx_event = self.tx_event.clone();
+            let mcp_pool = mcp_pool.clone();
+            tasks.push(async move {
+                let result = Engine::execute_tool_with_lock(
+                    lock,
+                    true,
+                    false,
+                    tx_event,
+                    tool_name.clone(),
+                    tool_input.clone(),
+                    Some(registry_ref),
+                    mcp_pool,
+                    None,
+                )
+                .await;
+                (tool_name, result)
+            });
+        }
+
+        let mut results = Vec::new();
+        while let Some((tool_name, result)) = tasks.next().await {
+            match result {
+                Ok(output) => {
+                    let mut error = None;
+                    if !output.success {
+                        error = Some(output.content.clone());
+                    }
+                    results.push(ParallelToolResultEntry {
+                        tool_name,
+                        success: output.success,
+                        content: output.content,
+                        error,
+                    });
+                }
+                Err(err) => {
+                    let message = format!("{err}");
+                    results.push(ParallelToolResultEntry {
+                        tool_name,
+                        success: false,
+                        content: format!("Error: {message}"),
+                        error: Some(message),
+                    });
+                }
+            }
+        }
+
+        ToolResult::json(&ParallelToolResult { results })
+            .map_err(|e| ToolError::execution_failed(e.to_string()))
+    }
+
+    #[allow(clippy::too_many_arguments)]
+    pub(super) async fn execute_tool_with_lock(
+        lock: Arc<RwLock<()>>,
+        supports_parallel: bool,
+        interactive: bool,
+        tx_event: mpsc::Sender<Event>,
+        tool_name: String,
+        tool_input: serde_json::Value,
+        registry: Option<&crate::tools::ToolRegistry>,
+        mcp_pool: Option<Arc<AsyncMutex<McpPool>>>,
+        context_override: Option<crate::tools::ToolContext>,
+    ) -> Result<ToolResult, ToolError> {
+        let _guard = if supports_parallel {
+            ToolExecGuard::Read(lock.read().await)
+        } else {
+            ToolExecGuard::Write(lock.write().await)
+        };
+
+        if interactive {
+            let _ = tx_event.send(Event::PauseEvents).await;
+        }
+
+        let result = if McpPool::is_mcp_tool(&tool_name) {
+            if let Some(pool) = mcp_pool {
+                Engine::execute_mcp_tool_with_pool(pool, &tool_name, tool_input).await
+            } else {
+                Err(ToolError::not_available(format!(
+                    "tool '{tool_name}' is not registered"
+                )))
+            }
+        } else if let Some(registry) = registry {
+            registry
+                .execute_full_with_context(&tool_name, tool_input, context_override.as_ref())
+                .await
+        } else {
+            Err(ToolError::not_available(format!(
+                "tool '{tool_name}' is not registered"
+            )))
+        };
+
+        if interactive {
+            let _ = tx_event.send(Event::ResumeEvents).await;
+        }
+
+        result
+    }
+}
